Task 1

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -7,7 +7,7 @@
  *
  * Return: pointer to destination buffer @dest
  */
-void *_memcpy(void *dest, const void *src, size_t n)
+void *_memcpy(void *dest, const void *src, unsigned int n)
 {
 	unsigned int i;
